requirements.txt
openmeteo_requests==1.5.0
pandas==2.2.3
psycopg2==2.9.10
python-dotenv==1.1.0
requests-cache==1.2.1
retry-requests==2.0.0
SQLAlchemy==2.0.41